Resolve merge mismatch
